requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
pdf2image
youtube_transcript_api
pathlib
webvtt-py
langchain-community
langchain_openai
requests